Migrate NeoPixelTree firmware
One of the things I wanted to add to the original LedEyes project was a
NeoPixel ring from Adafruit to act as a candle inside of the pumpkin to
provide light for the mouth and nose of the Jack-O-Lantern. I had
previously worked on a candle flicker animation for my
https://github.com/adamgreen/NeoPixelTree project so this commit pulls
that code over. I will use it as a starting point for this project.

This commit also:
* Pulls in the gcc4mbed submodule that the firmware utilizes.
* Added .gitignore from NeoPixelTree project to ignore build output.