Refactor pupil animation
I implemented a new class called PupilAnimation which should be
versatile enough to implement any of michaltj's effects which just
animate the location of the pupils. This includes crazy spin, crossed
eyes, meth eyes, and lazy eye. There is one other effect called glow
eyes which animates the intensity and not the pupil position so it
won't be able to take advantage of this class.

This new PupilAnimation takes an array of key frames as an input. Each
key frame gives the new desired location of both the right and left
pupil and the time delay to use between each subframe used to
animate the pupils to their new locations. The animation can jump
the pupils to their new location or interpolate them to their new
location. It also supports speeding up or slowing down the time
between interpolated subframes to simulate acceleration.

I have already simplified the implementation of the existing
MoveEyeAnimation to take advantage of this new class. The other
special effects will be ported later.

I also moved the ARRAY_SIZE macro out of main.cpp and placed it in
util.h instead so that it can be used by multiple modules.